Introduce Viz.svgXmlToPngBase64 and add callback API for Viz.svgXmlToPngImageElement.
Viz.svgXmlToPngBase64 calls the provided callback with the base64-encoded version of the PNG image representation of the SVG, or an error. The error argument is null if no error occurred.

  Viz.svgXmlToPngBase64(xml, 2, function(error, data) {
    if (error) {
      // Do something about error
    } else {
      // Do something with data
    }
  });

Viz.svgXmlToPngImageElement can now be passed a callback, which is called with either an image element or an error:

  Viz.svgXmlToPngImageElement(xml, 2, function(error, image) {
    if (error) {
      // Do something about error
    } else {
      // Do something with image
    }
  });